+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/access.rb
@@ -1,7 +1,11 @@
+require_relative "./node"
+
 module Ladb::OpenCutList
 module Dentaku
   module AST
-    class Access
+    class Access < Node
+      attr_reader :structure, :index
+
       def self.arity
         2
       end
@@ -34,6 +38,10 @@ module Dentaku
 
       def type
         nil
+      end
+
+      def accept(visitor)
+        visitor.visit_access(self)
       end
     end
   end