+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/case.rb
@@ -9,6 +9,8 @@ module Ladb::OpenCutList
 module Dentaku
   module AST
     class Case < Node
+      attr_reader :switch, :conditions, :else
+
       def self.min_param_count
         2
       end
@@ -56,6 +58,10 @@ module Dentaku
         switch_dependencies(context) +
         condition_dependencies(context) +
         else_dependencies(context)
+      end
+
+      def accept(visitor)
+        visitor.visit_case(self)
       end
 
       private