+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/case/case_when.rb
@@ -2,6 +2,8 @@ module Ladb::OpenCutList
 module Dentaku
   module AST
     class CaseWhen < Operation
+      attr_reader :node
+
       def initialize(node)
         @node = node
       end
@@ -24,6 +26,14 @@ module Dentaku
 
       def self.max_param_count
         1
+      end
+
+      def accept(visitor)
+        visitor.visit_when(self)
+      end
+
+      def to_s
+        'WHEN'
       end
     end
   end