+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/functions/filter.rb
@@ -3,13 +3,13 @@ require_relative './enum'
 module Ladb::OpenCutList
 module Dentaku
   module AST
-    class All < Enum
+    class Filter < Enum
       def value(context = {})
         collection      = Array(@args[0].value(context))
         item_identifier = @args[1].identifier
         expression      = @args[2]
 
-        collection.all? do |item_value|
+        collection.select do |item_value|
           expression.value(
             context.merge(
               FlatHash.from_hash_with_intermediates(item_identifier => item_value)
@@ -23,5 +23,5 @@ end
 end
 
 module Ladb::OpenCutList
-  Dentaku::AST::Function.register_class(:all, Dentaku::AST::All)
+  Dentaku::AST::Function.register_class(:filter, Dentaku::AST::Filter)
 end