+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/functions/map.rb
@@ -1,31 +1,18 @@
-require_relative '../function'
-require_relative '../../exceptions'
+require_relative './enum'
 
 module Ladb::OpenCutList
 module Dentaku
   module AST
-    class Map < Function
-      def self.min_param_count
-        3
-      end
-
-      def self.max_param_count
-        3
-      end
-
-      def deferred_args
-        [1, 2]
-      end
-
+    class Map < Enum
       def value(context = {})
-        collection      = @args[0].value(context)
+        collection      = Array(@args[0].value(context))
         item_identifier = @args[1].identifier
         expression      = @args[2]
 
         collection.map do |item_value|
           expression.value(
-            context.update(
-              FlatHash.from_hash(item_identifier => item_value)
+            context.merge(
+              FlatHash.from_hash_with_intermediates(item_identifier => item_value)
             )
           )
         end