+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/grouping.rb
@@ -1,7 +1,9 @@
+require_relative "./node"
+
 module Ladb::OpenCutList
 module Dentaku
   module AST
-    class Grouping
+    class Grouping < Node
       def initialize(node)
         @node = node
       end