+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/nil.rb
@@ -5,6 +5,10 @@ module Dentaku
       def value(*)
         nil
       end
+
+      def accept(visitor)
+        visitor.visit_nil(self)
+      end
     end
   end
 end