+++ src/ladb_opencutlist/ruby/lib/dentaku/ast/string.rb
@@ -4,6 +4,13 @@ module Ladb::OpenCutList
 module Dentaku
   module AST
     class String < Literal
+      def quoted
+        %Q{"#{ escaped }"}
+      end
+
+      def escaped
+        @value.gsub('"', '\"')
+      end
     end
   end
 end