+++ CHANGELOG.md
@@ -10,6 +10,7 @@ in 3.x, 2.x, 1.x and 0.x versions.
   * Added generate cutlist menu item
   * Added progress feedback on generate cutlist and list materials
   * Added page description to printed part list
+  * Added re-load material texture 
   * Improved ruby i18n string getter to support $()
   * Improved disabled edge cell display in part list
   * Fixed edge std size display