+++ CHANGELOG.md
@@ -16,6 +16,7 @@ in 3.x, 2.x, 1.x and 0.x versions.
   * Improved disabled edge cell display in part list
   * Fixed edge std size display
   * Remove URI.escape deprecation uses
+  * Now only standard sheet and bar uses dim prices in reports
 
 * 3.0.2 (2022-02-27)